task: Rename to common Module.
* rename path.
* rename module names.
