+++ src/common/enums/common-route.enum.ts
@@ -1,0 +1,3 @@
+export enum CommonRoute {
+  health = '/health',
+}
