+++ src/common/enums/index.ts
@@ -1,0 +1,2 @@
+export * from './common-route.enum';
+export * from './environment.enum';
