+++ src/common/exceptions/index.ts
