--- src/core/enums/core-route.enum.ts
@@ -1,3 +1,0 @@
-export enum CoreRoute {
-  health = '/core/health',
-}
