--- src/core/enums/index.ts
@@ -1,2 +1,0 @@
-export * from './core-route.enum';
-export * from './environment.enum';
